refactor: cleaning up types
