+++ src/components/MyClaimCreateView/MyClaimCreateView.tsx
@@ -11,7 +11,7 @@ import FeedbackService, { notifySuccess } from '../../services/FeedbackService'
 import * as Claims from '../../state/ducks/Claims'
 import * as Wallet from '../../state/ducks/Wallet'
 import { State as ReduxState } from '../../state/PersistentStore'
-import { ICType } from '../../types/Ctype'
+import { ICType, IClaimInput } from '../../types/Ctype'
 import { BlockUi } from '../../types/UserFeedback'
 import CTypeUtils from '../../services/CtypeUtils'
 
@@ -165,7 +165,7 @@ class MyClaimCreateView extends Component<Props, State> {
     this.setState({ name: e.target.value })
   }
 
-  private claimInputModel(cType: any, lang?: string): any {
+  private claimInputModel(cType: sdk.ICType, lang?: string): IClaimInput {
     return CTypeUtils.getClaimInputModel(cType, lang)
   }
 }
